eldarica: split horn converter from horn prover, deduplicate predicates
